django_rest/requirements.txt
Django
psycopg2
djangorestframework
pygments
django-cors-headers